Update patch set 1
Patch Set 1:

See the other half of the topic. I had written policy for this on 17.1 that never got picked to 18.1. Having the prop on vendor matches stock deadpool and allows devices without a product partition (like the shield android tv's) to set the prop where a gsi would see it.

Patch-set: 1