TVLauncherNoGMS: Allow launcher to run with gms installed
These copies of the apps refuse to run if gms or gsf is detected.
Use smali to replace instances of com.google.android.gms and .gsf
with misnamed strings, so the checks always fail.

Because the apks have been modified, the signature is broken. So
resign with the platform key. These apps cannot be updated off the
play store anyways.

Change-Id: I395a3254612045138ada8cc3c17b8b75eba433e0